feat: followups when channel already deleted

+++ onclick.js
@@ -7,6 +7,10 @@ exports.respond = async function (interaction) {
 	const channels = category.children;
 
 	const deleteChannel = channels.find((channel) => channel.name === interaction.customId);
+	if (!deleteChannel) {
+		await interaction.followUp({ content: interaction.customId + 'はすでに削除されています', ephemeral: true });
+		return;
+	}
 
 	await deleteChannel.delete();
 	await interaction.followUp({ content: interaction.customId + 'を削除しました', ephemeral: true });
